rename common to pass_manager
it's not really a pass manager, but common wasn't the right name either
