+++ src/functional/iterator/transforms/__init__.py
@@ -1,4 +1,4 @@
-from functional.iterator.transforms.common import apply_common_transforms
+from functional.iterator.transforms.pass_manager import apply_common_transforms
 
 
 __all__ = ["apply_common_transforms"]
